+++ packages/plugin-rsc/src/plugin.ts
@@ -174,7 +174,7 @@ export type RscPluginOptions = {
   /**
    * use `Plugin.buildApp` hook (introduced on Vite 7) instead of `builder.buildApp` configuration
    * for better composability with other plugins.
-   * @default false
+   * @default true since Vite 7
    */
   useBuildAppHook?: boolean
 
@@ -309,6 +309,14 @@ export default function vitePluginRsc(
     {
       name: 'rsc',
       async config(config, env) {
+        if (config.rsc) {
+          // mutate `rscPluginOptions` since internally this object is passed around
+          Object.assign(
+            rscPluginOptions,
+            // not sure which should win. for now plugin constructor wins.
+            vite.mergeConfig(config.rsc, rscPluginOptions),
+          )
+        }
         // crawl packages with "react" in "peerDependencies" to bundle react deps on server
         // see https://github.com/svitejs/vitefu/blob/d8d82fa121e3b2215ba437107093c77bde51b63b/src/index.js#L95-L101
         const result = await crawlFrameworkPkgs({
@@ -332,7 +340,7 @@ export default function vitePluginRsc(
         ]
 
         return {
-          appType: 'custom',
+          appType: config.appType ?? 'custom',
           define: {
             'import.meta.env.__vite_rsc_build__': JSON.stringify(
               env.command === 'build',
@@ -412,11 +420,26 @@ export default function vitePluginRsc(
           builder: {
             sharedPlugins: true,
             sharedConfigBuild: true,
-            buildApp: rscPluginOptions.useBuildAppHook ? undefined : buildApp,
+            async buildApp(builder) {
+              if (!rscPluginOptions.useBuildAppHook) {
+                await buildApp(builder)
+              }
+            },
           },
         }
       },
-      buildApp: rscPluginOptions.useBuildAppHook ? buildApp : undefined,
+      configResolved() {
+        if (Number(vite.version.split('.')[0]) >= 7) {
+          rscPluginOptions.useBuildAppHook ??= true
+        }
+      },
+      buildApp: {
+        async handler(builder) {
+          if (rscPluginOptions.useBuildAppHook) {
+            await buildApp(builder)
+          }
+        },
+      },
       configureServer(server) {
         ;(globalThis as any).__viteRscDevServer = server
 
@@ -989,9 +1012,10 @@ import.meta.hot.on("rsc:update", () => {
     ...vitePluginRscMinimal(rscPluginOptions, manager),
     ...vitePluginFindSourceMapURL(),
     ...vitePluginRscCss(rscPluginOptions, manager),
-    ...(rscPluginOptions.validateImports !== false
-      ? [validateImportPlugin()]
-      : []),
+    {
+      ...validateImportPlugin(),
+      apply: () => rscPluginOptions.validateImports !== false,
+    },
     scanBuildStripPlugin({ manager }),
     ...cjsModuleRunnerPlugin(),
     ...globalAsyncLocalStoragePlugin(),